requirements-test.txt
# Testing Dependencies for Crypto Trading Bot

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0          # Code coverage reporting
pytest-mock>=3.10.0        # Enhanced mocking capabilities
pytest-xdist>=3.0.0        # Parallel test execution

# HTTP mocking for API tests
responses>=0.23.0           # Mock HTTP requests/responses
requests-mock>=1.10.0      # Alternative HTTP mocking

# Test utilities
freezegun>=1.2.0           # Time/date mocking
factory-boy>=3.2.0         # Test data generation
faker>=18.0.0              # Fake data generation

# Assertion helpers
pytest-sugar>=0.9.0       # Better test output
pytest-clarity>=1.0.0     # Better assertion diffs

# Async testing (if needed later)
pytest-asyncio>=0.21.0

# Development tools
pre-commit>=3.0.0          # Code quality hooks